temp_backup/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.0
requests==2.31.0 